Entrega da atividade

+++ src/senaifit/services/AtividadeService.java
@@ -4,40 +4,51 @@ import java.util.Optional;
 
 import javax.validation.Valid;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
 
+import senaifit.DTO.AtividadeDTO;
 import senaifit.entities.Atividade;
 import senaifit.repositories.AtividadeRepository;
 
 @Service
 public class AtividadeService {
 
-    @Autowired
     private AtividadeRepository atividadeRepo;
 
-    public void salvaAtividade(@Valid Atividade atividade) {
+    public AtividadeService(AtividadeRepository atividadeRepo) {
+	this.atividadeRepo = atividadeRepo;
+    }
 
-	this.atividadeRepo.save(atividade);
+    public String cadastraAtividade(@Valid @RequestBody AtividadeDTO atividadeDTO) {
+
+	Atividade atividade = new Atividade();
+	atividade.setNome(atividadeDTO.getNome());
+
+	if (!(atividade.getNome().isEmpty() && atividade.equals(null))) {
+	    this.atividadeRepo.save(atividade);
+	    return "Atividade cadastrada com sucesso";
+	}
+	return "Erro ao cadastrar atividade";
     }
 
     public Optional<Atividade> obtemAtividade(long atividadeId) {
 
-	return this.atividadeRepo.findById(atividadeId);
+	Optional<Atividade> atividade = this.atividadeRepo.findById(atividadeId);
+
+	if (atividade.isPresent()) {
+	    return atividade;
+	}
+	return Optional.empty();
     }
 
-    public Optional<Atividade> obtemAtividadePorNome(String nomeAtividade) {
+    public boolean deletaAtividadePorId(long atividadeId) {
 
-	return this.atividadeRepo.findByNome(nomeAtividade);
-    }
-
-    public void deletaAtividade(Atividade atividade) {
-
-	this.atividadeRepo.delete(atividade);
-    }
-
-    public void deletaAtividadePorId(long atividadeId) {
-
-	this.atividadeRepo.deleteById(atividadeId);
+	Optional<Atividade> atividade = obtemAtividade(atividadeId);
+	if (atividade.isPresent()) {
+	    this.atividadeRepo.deleteById(atividadeId);
+	    return true;
+	}
+	return false;
     }
 }
